Update for fix in bills, splits
Fix bug when the end search results in cur_end having the query result.

Pages now have been fixed so that the last name and "acreage" are split,
so, the line number is now one more, and doesn't need to be split at
"acreage". Do still handle the case where acreage and the last name are
combined, however, but if it is empty, then pick the next line.